# flashcard-be/api-docs.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "REST API",
    "description": "REST API description...",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerToken": []
    }
  ],
...
  "paths": {
...
    "/api/v1/user": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get user details by ID",
        "operationId": "getUserById",
        "requestBody": {
          "description": "User ID for fetching user details",
          "content": {
            "application/json": {
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "User not found"
          },
          "200": {
            "description": "Successfully fetched user details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Create a new user",
        "operationId": "createUser",
        "requestBody": {
          "description": "New user registration request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Invalid user request"
          },
          "201": {
            "description": "User successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/user/login": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Authenticate and log in the user",
        "operationId": "login",
        "requestBody": {
          "description": "User login request with username and password",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "401": {
            "description": "Invalid login credentials"
          },
          "200": {
            "description": "User successfully logged in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/user/check-token": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Check the validity of a user token",
        "operationId": "checkToken",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid token"
          },
          "200": {
            "description": "Token is valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/user/{userId}": {
      "delete": {
        "tags": [
          "User"
        ],
        "summary": "Delete a user by ID",
        "operationId": "deleteUser",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User successfully deleted"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AppUserRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 50,
            "minLength": 3
          },
          "password": {
            "type": "string",
            "maxLength": 50,
            "minLength": 6
          }
        },
        "required": [
          "password",
          "username"
        ]
      },
      "AppUserResponse": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/AppUserResponse"
          },
          "token": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}